home loan regarding work

+++ loan-service/src/main/java/com/bms/loan/Repository/LoanRepository.java
@@ -1,12 +1,12 @@
 package com.bms.loan.Repository;
 
-import com.bms.loan.entity.Loans;
+import com.bms.loan.entity.loan.Loans;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface LoanRepository extends JpaRepository<Loans, Integer> {
+public interface LoanRepository extends JpaRepository<Loans, Long> {
     Optional<Loans> findByLoanId(Long loanApplicationId);
 }
